Add apply-operator `$`.
