+++ nodes/oper_nodes.py
@@ -10,6 +10,7 @@ from nodes.base_oper import *
 from nodes.expression import *
 from bases.ntype import *
 from lang import *
+from nodes.func_features import *
 from nodes.oper_dot import *
 from nodes.structs import StructConstrBegin 
 from vars import *
@@ -557,15 +558,6 @@ class OpBitwise(BinOper):
         # dprint('tpl % args: ', tpl % args)
         return StringVal(tpl % args)
 
-class UnarOper(OperCommand):
-    def __init__(self, oper:str, inner:Expression=None):
-        super().__init__(oper)
-        self.inner = inner
-        # self.res = None
-
-    def setInner(self, inner:Expression):
-        self.inner = inner
-
 
 class BoolNot(UnarOper):
     def __init__(self, oper:str, inner:Expression=None):
@@ -885,28 +877,4 @@ class RegexpSearchOper(RegexpOper):
         src = var2val(src)
         res = rx.find(src)
         self.res = res
-        
-from nodes.func_features import *
-class RTildArrowExpr(UnarOper):
-
-    def __init__(self, inner = None):
-        super().__init__('~>', inner)
-
-    # def setInner(self, inner:Expression):
-    #     self.inner = inner
-
-    def do(self, ctx:Context):
-        ''' '''
-        # do inner
-        self.inner.do(ctx)
-        # get func from inner
-        fval = self.inner.get()
-        func = var2val(fval)
-        # check func for currying
-        if func.getType() != TypeFunc():
-            raise EvalErr('Not function tried to curry')
-        # call curry(), get result
-        curFunc = func_curry(ctx, func)
-        self.res = curFunc
-
