+++ parser.py
@@ -26,7 +26,7 @@ c_regex = '| / % '
 
 opers = [n for n in ('; , ... $$ .. ** ++ -- += -= *= /= %=  && || == != <= >= << >> => ?> !?> -> <- !- := ?: /: !: :? :> :: =~ ?~ /~ ~>'
 # ' """ \'\'\' ``` '
-' < > = + - * / | \\ { } [ ] . , : ? ~ ! % ^ & * ( )').split(' ') if n]
+' < > = + - * / | \\ { } [ ] . , : ? ~ ! % ^ & * $ ( )').split(' ') if n]
 
 # if i > 0 and el.text in ['-', '+', '!', '~'] and elems[i-1].type == Lt.oper and elems[i-1].text != ')'
 unarOpers = '- + ! ~'
